refactor: post condition asset info
