+++ .changeset/nice-humans-chew.md
@@ -1,0 +1,5 @@
+---
+'@stacks/wallet-web': patch
+---
+
+This fixes how asset metadata displays in post conditions. It should display correctly now even if a user doesn't have that token in their account.
